DSS-2881: Java - Save Loading Information API v1

+++ api/cp-dss-api/gateway/src/main/java/com/cpdss/gateway/domain/loadingplan/ToppingOffSequence.java
@@ -9,7 +9,7 @@ public class ToppingOffSequence {
 
   private Long id;
   private Long loadingInfoId;
-  private Long orderNumber;
+  private Integer orderNumber;
   private Long tankId;
   private Long cargoId;
   private String shortName;
